pyproject.toml
[project]
name = "rm-gen"
version = "0.1.0"
description = "RM-Gen, a radiomap dataset generator"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    { name = "Artur Padovesi Piratelli" },
]
requires-python = ">=3.11"
keywords = ["pipeline", "raytracing", "ray tracing", "dataset", "wireless communication", "radio map"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Telecommunications Industry",
    "Operating System :: OS Independent"
]

dependencies = [
    "geopandas>=1.1.1",
    "mapbox-earcut>=2.0.0",
    "marshmallow>=4.1.1",
    "matplotlib>=3.10.7",
    "opencv-python>=4.12.0.88",
    "osmium>=4.2.0",
    "osmnx>=2.0.7",
    "rasterio>=1.4.3",
    "requests>=2.32.5",
    "scipy>=1.16.3",
    "shapely>=2.1.2",
    "sionna-rt>=1.2.1",
    "tqdm>=4.67.1",
    "trimesh>=4.10.1",
]


[project.urls]
source = "https://github.com/artistrea/rm-gen/"
# documentation = "https://nvlabs.github.io/sionna/"
# issues = "https://github.com/nvlabs/sionna/issues"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
examples = [
    "notebook>=7.5.0",
]

doc = [
    "docutils==0.21.2",
    "markupsafe==3.0.2",
    "nbconvert==7.13.1",
    "nbsphinx==0.9.6",
    "sphinx==8.1.3",
    "sphinx-autodoc-typehints==2.5.0",
    "sphinx-copybutton==0.5.2",
    "sphinx-rtd-theme==3.0.2",
]

dev = [
    "pylint>=3.3.2"
]
